+++ src/services/Guide.php
@@ -10,16 +10,14 @@
 
 namespace wbrowar\guide\services;
 
+use Craft;
+use craft\base\Component;
 use craft\elements\User;
 use wbrowar\guide\Guide as GuidePlugin;
 use wbrowar\guide\helpers\GuideHelpers;
 use wbrowar\guide\models\Guide as GuideModel;
-use wbrowar\guide\records\Placements;
-
-use Craft;
-use craft\base\Component;
-use LitEmoji\LitEmoji;
 use wbrowar\guide\records\Guides;
+use wbrowar\guide\records\Placements;
 
 /**
  * @author    Will Browar
@@ -72,7 +70,6 @@ class Guide extends Component
                         $guide->content = GuideHelpers::decodeEmoji($guide->content);
                     }
                 }
-
                 break;
             case 'new':
                 $guides = new Guides([]);
@@ -86,7 +83,6 @@ class Guide extends Component
                 if (!empty($guides->content)) {
                     $guides->content = GuideHelpers::decodeEmoji($guides->content);
                 }
-
                 break;
             case 'count':
                 $guides = Guides::find()->where($params)->count();
@@ -166,8 +162,11 @@ class Guide extends Component
 
         $record->authorId = $model->authorId;
         $record->content = GuideHelpers::encodeEmoji($model->content);
+        $record->contentCss = $model->contentCss;
+        $record->contentJavascript = $model->contentJavascript;
         $record->contentSource = $model->contentSource;
         $record->contentUrl = $model->contentUrl;
+        $record->renderMarkdown = $model->renderMarkdown;
         $record->summary = $model->summary;
         $record->slug = $this->getUniqueSlug($model->slug, $id ?? 0);
         $record->template = $model->template;